+++ packages/botonic-react/src/webchat/context/index.tsx
@@ -13,6 +13,9 @@ export const WebchatContext = createContext<WebchatContextProps>({
   getThemeProperty: () => {
     return
   }, // used to retrieve a specific property of the theme defined by the developer in his 'webchat/index.js'
+  closeWebview: async () => {
+    return
+  },
   openWebview: () => {
     return
   },
